+++ c2c-plugin.php
@@ -48,6 +48,7 @@ abstract class c2c_Plugin_064 {
 		'input'            => '',
 		'input_attributes' => '',
 		'label'            => '',
+		'more_help'        => '',
 		'no_wrap'          => false,
 		'numbered'         => false,
 		'options'          => '',
@@ -1031,15 +1032,18 @@ HTML;
 			echo '</fieldset>';
 		} elseif ( $input == 'checkbox' ) {
 			echo "<input type='{$input}' {$attribs} value='1' " . checked( $value, 1, false ) . " />\n";
+			if ( ! empty( $this->config[ $opt ]['help'] ) ) {
+				printf( "<label class='description' for='%s'>%s</label>\n", $opt, $this->config[ $opt ]['help'] );
+				$this->config[ $opt ]['help'] = '';
+			}
 		} else { // Only 'text' and 'password' should fall through to here.
 			echo "<input type='{$input}' {$attribs} value='" . esc_attr( $value ) . "' />\n";
 		}
-		if ( $help = apply_filters( $this->get_hook( 'option_help'), $this->config[ $opt ]['help'], $opt ) ) {
-			if ( 'checkbox' === $input ) {
-				echo "<label class='description' for='{$opt}'>{$help}</label>\n";
-			} else {
-				echo "<p class='description'>{$help}</p>\n";
-			}
+		if ( $help = apply_filters( $this->get_hook( 'option_help'), $this->config[ $opt ]['help'], $opt, 'help' ) ) {
+			echo "<p class='description'>{$help}</p>\n";
+		}
+		if ( $help = apply_filters( $this->get_hook( 'option_help'), $this->config[ $opt ]['more_help'], $opt, 'more_help' ) ) {
+			echo "<p class='description'>{$help}</p>\n";
 		}
 
 		do_action( $this->get_hook( 'post_display_option' ), $opt );